factor out shared trajectory animation code
